Coin разделен на CmcCoin и RoomCoin
