+++ app/build.gradle
@@ -71,6 +71,9 @@ dependencies {
     implementation 'com.google.dagger:dagger:2.25.2'
     annotationProcessor 'com.google.dagger:dagger-compiler:2.25.2'
 
+    implementation "androidx.room:room-runtime:2.2.5"
+    annotationProcessor "androidx.room:room-compiler:2.2.5"
+
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
