+++ app/src/main/java/com/scorp/loftcoin/data/Listings.java
@@ -7,5 +7,5 @@ import java.util.List;
 @AutoValue
 abstract class Listings {
 
-    abstract List<AutoValue_Coin> data();
+    abstract List<AutoValue_CmcCoin> data();
 }
